Again reordered and summarized code (1); introduced restart on time out (2); player score set to zero (3)
(1): Summarized steps on level load and start into better readable member methods/ got rid of redundant code; restructured steps into more logical order.
(2): If seconds since level start reach zero on the game timer (see also there) the current level gets reloaded (number of kittens and objects stay the same). Introduced the boolean 'restartOnTimeOut' to control this behaviour: Furniture and cat tray objects won't be removed from scene and not re-instantiated.
(3): Instead of checking whether the score is greater than the minimal size of an integer the simpler and more classical approach is applied: checking if it is above zero; if not it is set to/ stays zero.